Add proof-of-concept Execute
This skips out on a lot of things like validation, arguments, aliases,
and more sophisticated types. However, this gives enough framing to
guide end-to-end tests that call into an object to get results.

There's also plenty of opportunities for optimization: reflection is
done lazily and on every request rather than all at once and checked
against the query type. However, the API is structured in such a way
that these improvements can be added incrementally with little to no
change in user experience.

No idea whether this is the final API, but it feels pretty good so far.