graphql: add ValidatedQuery type
The primary usage within the library is to be able to check during the
HTTP request parsing to prevent mutations from being sent in GET
requests. This only needs the AST to implement the check, but we want to
avoid parsing again when we Execute the query. Unfortunately, sending
the AST around isn't an operation that any typical user of the library
would want to do, so committing extra API surface to it would not be
worth its weight.

Instead, this change combines parsing and validation into a
`*Schema.Validate` method and permits storing the query into a new
optional `Request.ValidatedQuery` field. This gives the application
control over the validation process and permits caching if they so
choose. This may cause more work in `graphqlhttp.Parse`, but reduces
work for the overall request flow.

Removed confusing `Request.IsQuery()` method.